Change chat reference from integer to bigint

+++ priv/repo/migrations/20200810111457_create_chats.exs
@@ -5,7 +5,7 @@ defmodule Talbot.Repo.Migrations.CreateChats do
     create table(:chats, primary_key: false) do
       add :id, :uuid, primary_key: true
 
-      add :chat_reference, :integer
+      add :chat_reference, :bigint
 
       add :passcode, :string
       add :passcode_expires_at, :naive_datetime
